Refactor code #2
Change struct in order to delete unecessary double loop.
It was not known (to me) that we could have a struct without having to do a serialization or something like that to use in the function **write**. So the code was somehow "duplicate".
By changing the fields of the struct, we notice that the code works as desired, but without unnecessary repetition.